% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{verbose}
\alias{verbose}
\title{Set/Get the verbosity level}
\usage{
verbose(name = NULL, ...)
}
\arguments{
\item{name}{name for which verbosity level should be returned. Returns NULL
if no verbosity level is set for that name.}

\item{...}{key = value pairs to be set}
}
\description{
If no arguments are given, all current verbosity level are returned.
If a name is given, the verbosity level for that name is returned.
If no name is given and key = value pairs are provided
the verbosity levels for these keys will be set using the given values.
}
